[MSFT] Factor out Quartus/Stratix10 logic from TclEmitter
Refactoring commit which makes `TclEmitter` generic by allowing inheriting classes to provide custom `TclOutputState` classes. Furthermore, adds new arguments to `msft-export-tcl` to select tcl flavor and target device, respectively.
The intention of this refactor is to write TCL emitters for lowering MSFT to alternative flows and devices.
