+++ include/circt/InitAllTranslations.h
@@ -15,7 +15,6 @@
 #include "circt/Dialect/ESI/ESIDialect.h"
 #include "circt/Dialect/FIRRTL/FIREmitter.h"
 #include "circt/Dialect/FIRRTL/FIRParser.h"
-#include "circt/Dialect/MSFT/ExportTcl.h"
 #include "circt/Target/ExportSystemC.h"
 
 #ifndef CIRCT_INITALLTRANSLATIONS_H
