+++ lib/Dialect/MSFT/CMakeLists.txt
@@ -12,7 +12,9 @@ add_circt_dialect_library(CIRCTMSFT
   MSFTDialect.cpp
   MSFTOpInterfaces.cpp
   MSFTOps.cpp
-  ExportQuartusTcl.cpp
+  MSFTPasses.cpp
+  TclEmitter.cpp
+  TclEmitters/Quartus/TclEmitter.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${CIRCT_MAIN_INCLUDE_DIR}/circt/Dialect/MSFT
